Add Introduction section to README.md
Added a new "Introduction" section to the README.md file, including a welcome message and a brief description of the project. The description outlines that the project is a curated collection of hands-on experiments aimed at helping newcomers understand the basics of MonoGame. A placeholder text "More to follow." was also added at the end.